File management
All the files have been moved to their appropriate location.
