--- Src/Bin/pathQuick.json
@@ -1,180 +1,0 @@
-{
-    "path": [
-        1,
-        2652,
-        2644,
-        2651,
-        2649,
-        2643,
-        717,
-        2645,
-        2653,
-        2608,
-        2626,
-        2606,
-        2599,
-        2623,
-        2614,
-        2610,
-        2612,
-        2613,
-        2617,
-        2618,
-        2619,
-        2620,
-        2621,
-        2922,
-        2519,
-        2921,
-        2906,
-        2905,
-        2896,
-        2825,
-        2822,
-        2813,
-        2810,
-        2469,
-        2800,
-        2798,
-        2795,
-        2444,
-        2441,
-        2439,
-        2350,
-        2333,
-        2337,
-        2349,
-        2348,
-        2334,
-        2321,
-        2316,
-        2319,
-        2318,
-        2265,
-        2263,
-        2241,
-        2242,
-        3958,
-        4221,
-        4362,
-        4363,
-        4573,
-        4361,
-        1044,
-        4555,
-        4216,
-        4215,
-        4214,
-        3592,
-        2408,
-        2409,
-        1273,
-        3961,
-        3962,
-        4122,
-        4123,
-        4129,
-        4132,
-        4185,
-        4377,
-        426,
-        3963,
-        3964,
-        4697,
-        3965,
-        3966,
-        1467,
-        1570,
-        1337,
-        1334,
-        1478,
-        1585,
-        1596,
-        1460,
-        1595,
-        1635,
-        1320,
-        1594,
-        4665,
-        1262,
-        1279,
-        1286,
-        1287,
-        4531,
-        1264,
-        1120,
-        1159,
-        1214,
-        1212,
-        1278,
-        1307,
-        1199,
-        1277,
-        1306,
-        1280,
-        1281,
-        4388,
-        4518,
-        714,
-        715,
-        1128,
-        1276,
-        1167,
-        1275,
-        1297,
-        182,
-        266,
-        267,
-        116,
-        94,
-        95,
-        87,
-        88,
-        83,
-        93,
-        4386,
-        82136,
-        82134,
-        82135,
-        82164,
-        82163,
-        82033,
-        82133,
-        82038,
-        82132,
-        82081,
-        82130,
-        82131,
-        82078,
-        82128,
-        82142,
-        82141,
-        82083,
-        82082,
-        82030,
-        82029,
-        82028,
-        82026,
-        82027,
-        82025,
-        82002,
-        81997,
-        81996,
-        81959,
-        81958,
-        81894,
-        81893,
-        81896,
-        81971,
-        81898,
-        81897,
-        81983,
-        32,
-        31,
-        22,
-        27,
-        24,
-        5
-    ],
-    "time": 956027
-}
