--- Src/Bin/pathQuick.xml
@@ -1,4 +1,0 @@
-<Path>
-    <Landmark>1</Landmark>
-    <Time>0</Time>
-</Path>
