+++ Src/Client/Bin/pathQuick.json
@@ -1,0 +1,7 @@
+{
+    "path": [
+        1,
+        2
+    ],
+    "time": 5062
+}
