+++ Src/Client/Bin/pathQuick.xml
@@ -1,0 +1,5 @@
+<Path>
+    <Landmark>1</Landmark>
+    <Landmark>2</Landmark>
+    <Time>5062</Time>
+</Path>
