+++ Src/Client/startClient.sh
@@ -1,0 +1,3 @@
+clear
+g++ -std=c++17 main.cpp ../Libraries/Tinyxml2/tinyxml2.cpp -o Bin/pathQuick 
+./Bin/pathQuick
