+++ Src/pathQuickLauncher.sh
@@ -1,0 +1,2 @@
+sudo bash Server/startServer.sh &
+sudo bash Client/startClient.sh
