--- Src/start.sh
@@ -1,3 +1,0 @@
-clear
-g++ -std=c++17 main.cpp Libraries/Tinyxml2/tinyxml2.cpp -o Bin/pathQuick 
-./Bin/pathQuick
