--- Src/startServer.sh
@@ -1,3 +1,0 @@
-g++ -std=c++17 Api/server.cpp Libraries/Tinyxml2/tinyxml2.cpp -o Bin/server
-clear
-./Bin/server
